requirements.txt
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.2.0
html5lib==1.1
idna==3.4
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.5
jupyter_core==5.0.0
lxml==4.9.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.3
pluggy==1.0.0
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.1
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
tweepy==3.10.0
python-dotenv==0.17.1
